tests/submission/Microsoft/semantic/nav.html

<!DOCTYPE html>
<html><head>
  <meta content="text/html; charset=utf-8" http-equiv="content-type">
  <title>verify NAV is a HTMLElement</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 <link href="http://www.w3.org/TR/html5/the-nav-element.html#the-nav-element" rel="help">
</head>
<body>
<div id="log"></div>
<nav id="nav">
 <h1>Navigation</h1>
 <ul>
  <li><a href="#">test link 1</a></li>
  <li><a href="#">test link 2</a></li>
 </ul>
</nav>
<script type="text/javascript">
 test(function()
 {
    assert_array_equals( "[object HTMLElement]",
                         document.getElementById("nav").toString() )
 })
</script>


</body></html>
